providers(github): Don't try to split None.
I've tried, it doesn't work.

When a github user such as I doesn't have a set name,
the code failed because the 'name' key exists and is set to None.

As we want to be sure we have a string here, let's use
`dict.get('name') or ""`

I've reset my name to blank on github as a test case :)